+++ src/Console/Command/ScheduleRunCommand.php
@@ -166,7 +166,7 @@ class ScheduleRunCommand extends Command
         return 0;
     }
 
-    /** @param iterable|array $tasks */
+    /** @param \SplFileInfo[] $tasks */
     private function fallbackTaskSource($tasks)
     {
         $tasksCount = \count($tasks);