+++ src/Filesystem/FilesystemInterface.php
@@ -42,4 +42,11 @@ interface FilesystemInterface
 
     /** @return string */
     public function projectRootDirectory();
+
+    /**
+     * @param string $filePath
+     *
+     * @return string
+     */
+    public function readContent($filePath);
 }