+++ src/Mailer.php
@@ -6,11 +6,7 @@ use Crunz\Configuration\Configuration;
 
 class Mailer
 {
-    /**
-     * Mailer instance.
-     *
-     * @param \Swift_Mailer
-     */
+    /** @var \Swift_Mailer|null */
     protected $mailer;
     /** @var Configuration */
     private $configuration;
@@ -130,14 +126,16 @@ class Mailer
     protected function getMessage($subject, $message)
     {
         $object = method_exists(\Swift_Message::class, 'newInstance')
-            ? \Swift_Message::newInstance()
-            : new \Swift_Message();
+            ? \Swift_Message::newInstance($subject, $message)
+            : new \Swift_Message($subject, $message)
+        ;
 
-        return  $object
-                 ->setBody($message)
-                 ->setSubject($subject)
-                 ->setFrom([$this->config('mailer.sender_email') => $this->config('mailer.sender_name')])
-                 ->setTo($this->config('mailer.recipients'));
+        $object
+            ->setFrom([$this->config('mailer.sender_email') => $this->config('mailer.sender_name')])
+            ->setTo($this->config('mailer.recipients'))
+        ;
+
+        return $object;
     }
 
     private function config($key)