+++ tests/TestCase/EndToEnd/Environment/Environment.php
@@ -24,7 +24,6 @@ final class Environment
     private $envFlags;
 
     /**
-     * @param string   $name
      * @param string[] $tasks
      *
      * @throws \Exception
@@ -220,6 +219,10 @@ final class Environment
             ],
             JSON_PRETTY_PRINT
         );
+
+        if (false === $content) {
+            throw new \RuntimeException("Unable to encode 'composer.json' content.");
+        }
 
         $this->filesystem
             ->dumpFile($composerJson->toString(), $content);