+++ tests/TestCase/EndToEnd/Environment/EnvironmentBuilder.php
@@ -19,7 +19,6 @@ final class EnvironmentBuilder
     /** @var EnvFlags */
     private $envFlags;
 
-    /** @param string $name */
     public function __construct(FilesystemInterface $filesystem, EnvFlags $envFlags)
     {
         $this->taskDirectory = Path::fromStrings('tasks');